Move methods from DatasetRecordStorage to its manager.
This is a mostly-mechanical change; for each method:

- move a method to the manager class;

- give it a dataset type argument, and have it look up the storage
  object internally;

- change calling code (SqlRegistry) to stop looking up the storage
  object before calling it.

The ultimate goal is to get rid of DimensionRecordStorageManager, at
least as a public interface.  But the steps after this one won't be
nearly as mechanical, so it's useful to separate them.
