+++ python/lsst/daf/butler/direct_query_driver/_driver.py
@@ -1133,7 +1133,6 @@ class DirectQueryDriver(QueryDriver):
         fields : `~collections.abc.Set` [ `str` ]
             Dataset fields to include.
         """
-        storage = self.managers.datasets[resolved_search.name]
         # The next two asserts will need to be dropped (and the implications
         # dealt with instead) if materializations start having dataset fields.
         assert (
@@ -1142,7 +1141,11 @@ class DirectQueryDriver(QueryDriver):
         assert (
             resolved_search.name not in joiner.timespans
         ), "Dataset timespan has unexpectedly already been joined in."
-        joiner.join(storage.make_query_joiner(resolved_search.collection_records, fields))
+        joiner.join(
+            self.managers.datasets.make_query_joiner(
+                self.get_dataset_type(resolved_search.name), resolved_search.collection_records, fields
+            )
+        )
 
 
 @dataclasses.dataclass
