+++ python/lsst/daf/butler/registry/interfaces/_datasets.py
@@ -32,7 +32,7 @@ from ... import ddl
 __all__ = ("DatasetRecordStorageManager", "DatasetRecordStorage")
 
 from abc import ABC, abstractmethod
-from collections.abc import Iterable, Iterator, Mapping, Sequence, Set
+from collections.abc import Iterable, Mapping, Sequence, Set
 from typing import TYPE_CHECKING, Any
 
 from lsst.daf.relation import Relation
@@ -66,289 +66,6 @@ class DatasetRecordStorage(ABC):
 
     def __init__(self, datasetType: DatasetType):
         self.datasetType = datasetType
-
-    @abstractmethod
-    def insert(
-        self,
-        run: RunRecord,
-        dataIds: Iterable[DataCoordinate],
-        idGenerationMode: DatasetIdGenEnum = DatasetIdGenEnum.UNIQUE,
-    ) -> Iterator[DatasetRef]:
-        """Insert one or more dataset entries into the database.
-
-        Parameters
-        ----------
-        run : `RunRecord`
-            The record object describing the `~CollectionType.RUN` collection
-            this dataset will be associated with.
-        dataIds : `~collections.abc.Iterable` [ `DataCoordinate` ]
-            Expanded data IDs (`DataCoordinate` instances) for the
-            datasets to be added.   The dimensions of all data IDs must be the
-            same as ``self.datasetType.dimensions``.
-        idGenerationMode : `DatasetIdGenEnum`
-            With `UNIQUE` each new dataset is inserted with its new unique ID.
-            With non-`UNIQUE` mode ID is computed from some combination of
-            dataset type, dataId, and run collection name; if the same ID is
-            already in the database then new record is not inserted.
-
-        Returns
-        -------
-        datasets : `~collections.abc.Iterable` [ `DatasetRef` ]
-            References to the inserted datasets.
-        """
-        raise NotImplementedError()
-
-    @abstractmethod
-    def import_(
-        self,
-        run: RunRecord,
-        datasets: Iterable[DatasetRef],
-    ) -> Iterator[DatasetRef]:
-        """Insert one or more dataset entries into the database.
-
-        Parameters
-        ----------
-        run : `RunRecord`
-            The record object describing the `~CollectionType.RUN` collection
-            this dataset will be associated with.
-        datasets : `~collections.abc.Iterable` of `DatasetRef`
-            Datasets to be inserted.  Datasets can specify ``id`` attribute
-            which will be used for inserted datasets. All dataset IDs must
-            have the same type (`int` or `uuid.UUID`), if type of dataset IDs
-            does not match type supported by this class then IDs will be
-            ignored and new IDs will be generated by backend.
-
-        Returns
-        -------
-        datasets : `~collections.abc.Iterable` [ `DatasetRef` ]
-            References to the inserted or existing datasets.
-
-        Notes
-        -----
-        The ``datasetType`` and ``run`` attributes of datasets are supposed to
-        be identical across all datasets but this is not checked and it should
-        be enforced by higher level registry code. This method does not need
-        to use those attributes from datasets, only ``dataId`` and ``id`` are
-        relevant.
-        """
-        raise NotImplementedError()
-
-    @abstractmethod
-    def delete(self, datasets: Iterable[DatasetRef]) -> None:
-        """Fully delete the given datasets from the registry.
-
-        Parameters
-        ----------
-        datasets : `~collections.abc.Iterable` [ `DatasetRef` ]
-            Datasets to be deleted.  All datasets must be resolved and have
-            the same `DatasetType` as ``self``.
-
-        Raises
-        ------
-        AmbiguousDatasetError
-            Raised if any of the given `DatasetRef` instances is unresolved.
-        """
-        raise NotImplementedError()
-
-    @abstractmethod
-    def associate(self, collection: CollectionRecord, datasets: Iterable[DatasetRef]) -> None:
-        """Associate one or more datasets with a collection.
-
-        Parameters
-        ----------
-        collection : `CollectionRecord`
-            The record object describing the collection.  ``collection.type``
-            must be `~CollectionType.TAGGED`.
-        datasets : `~collections.abc.Iterable` [ `DatasetRef` ]
-            Datasets to be associated.  All datasets must be resolved and have
-            the same `DatasetType` as ``self``.
-
-        Raises
-        ------
-        AmbiguousDatasetError
-            Raised if any of the given `DatasetRef` instances is unresolved.
-
-        Notes
-        -----
-        Associating a dataset with into collection that already contains a
-        different dataset with the same `DatasetType` and data ID will remove
-        the existing dataset from that collection.
-
-        Associating the same dataset into a collection multiple times is a
-        no-op, but is still not permitted on read-only databases.
-        """
-        raise NotImplementedError()
-
-    @abstractmethod
-    def disassociate(self, collection: CollectionRecord, datasets: Iterable[DatasetRef]) -> None:
-        """Remove one or more datasets from a collection.
-
-        Parameters
-        ----------
-        collection : `CollectionRecord`
-            The record object describing the collection.  ``collection.type``
-            must be `~CollectionType.TAGGED`.
-        datasets : `~collections.abc.Iterable` [ `DatasetRef` ]
-            Datasets to be disassociated.  All datasets must be resolved and
-            have the same `DatasetType` as ``self``.
-
-        Raises
-        ------
-        AmbiguousDatasetError
-            Raised if any of the given `DatasetRef` instances is unresolved.
-        """
-        raise NotImplementedError()
-
-    @abstractmethod
-    def certify(
-        self,
-        collection: CollectionRecord,
-        datasets: Iterable[DatasetRef],
-        timespan: Timespan,
-        context: SqlQueryContext,
-    ) -> None:
-        """Associate one or more datasets with a calibration collection and a
-        validity range within it.
-
-        Parameters
-        ----------
-        collection : `CollectionRecord`
-            The record object describing the collection.  ``collection.type``
-            must be `~CollectionType.CALIBRATION`.
-        datasets : `~collections.abc.Iterable` [ `DatasetRef` ]
-            Datasets to be associated.  All datasets must be resolved and have
-            the same `DatasetType` as ``self``.
-        timespan : `Timespan`
-            The validity range for these datasets within the collection.
-        context : `SqlQueryContext`
-            The object that manages database connections, temporary tables and
-            relation engines for this query.
-
-        Raises
-        ------
-        AmbiguousDatasetError
-            Raised if any of the given `DatasetRef` instances is unresolved.
-        ConflictingDefinitionError
-            Raised if the collection already contains a different dataset with
-            the same `DatasetType` and data ID and an overlapping validity
-            range.
-        CollectionTypeError
-            Raised if
-            ``collection.type is not CollectionType.CALIBRATION`` or if
-            ``self.datasetType.isCalibration() is False``.
-        """
-        raise NotImplementedError()
-
-    @abstractmethod
-    def decertify(
-        self,
-        collection: CollectionRecord,
-        timespan: Timespan,
-        *,
-        dataIds: Iterable[DataCoordinate] | None = None,
-        context: SqlQueryContext,
-    ) -> None:
-        """Remove or adjust datasets to clear a validity range within a
-        calibration collection.
-
-        Parameters
-        ----------
-        collection : `CollectionRecord`
-            The record object describing the collection.  ``collection.type``
-            must be `~CollectionType.CALIBRATION`.
-        timespan : `Timespan`
-            The validity range to remove datasets from within the collection.
-            Datasets that overlap this range but are not contained by it will
-            have their validity ranges adjusted to not overlap it, which may
-            split a single dataset validity range into two.
-        dataIds : `~collections.abc.Iterable` [ `DataCoordinate` ], optional
-            Data IDs that should be decertified within the given validity range
-            If `None`, all data IDs for ``self.datasetType`` will be
-            decertified.
-        context : `SqlQueryContext`
-            The object that manages database connections, temporary tables and
-            relation engines for this query.
-
-        Raises
-        ------
-        CollectionTypeError
-            Raised if ``collection.type is not CollectionType.CALIBRATION``.
-        """
-        raise NotImplementedError()
-
-    @abstractmethod
-    def make_relation(
-        self,
-        *collections: CollectionRecord,
-        columns: Set[str],
-        context: SqlQueryContext,
-    ) -> Relation:
-        """Return a `sql.Relation` that represents a query for for this
-        `DatasetType` in one or more collections.
-
-        Parameters
-        ----------
-        *collections : `CollectionRecord`
-            The record object(s) describing the collection(s) to query.  May
-            not be of type `CollectionType.CHAINED`.  If multiple collections
-            are passed, the query will search all of them in an unspecified
-            order, and all collections must have the same type.  Must include
-            at least one collection.
-        columns : `~collections.abc.Set` [ `str` ]
-            Columns to include in the relation.  See `Query.find_datasets` for
-            most options, but this method supports one more:
-
-            - ``rank``: a calculated integer column holding the index of the
-              collection the dataset was found in, within the ``collections``
-              sequence given.
-        context : `SqlQueryContext`
-            The object that manages database connections, temporary tables and
-            relation engines for this query.
-
-        Returns
-        -------
-        relation : `~lsst.daf.relation.Relation`
-            Representation of the query.
-        """
-        raise NotImplementedError()
-
-    @abstractmethod
-    def make_query_joiner(self, collections: Sequence[CollectionRecord], fields: Set[str]) -> QueryJoiner:
-        """Make a `..direct_query_driver.QueryJoiner` that represents a search
-        for datasets of this type.
-
-        Parameters
-        ----------
-        collections : `~collections.abc.Sequence` [ `CollectionRecord` ]
-            Collections to search, in order, after filtering out collections
-            with no datasets of this type via collection summaries.
-        fields : `~collections.abc.Set` [ `str` ]
-            Names of fields to make available in the joiner.  Options include:
-
-            - ``dataset_id`` (UUID)
-            - ``run`` (collection name, `str`)
-            - ``collection`` (collection name, `str`)
-            - ``collection_key`` (collection primary key, manager-dependent)
-            - ``timespan`` (validity range, or unbounded for non-calibrations)
-            - ``ingest_date`` (time dataset was ingested into repository)
-
-            Dimension keys for the dataset type's required dimensions are
-            always included.
-
-        Returns
-        -------
-        joiner : `..direct_query_driver.QueryJoiner`
-            A query-construction object representing a table or subquery.  If
-            ``fields`` is empty or ``len(collections) <= 1``, this is
-            guaranteed to have rows that are unique over dimension keys.
-        """
-        raise NotImplementedError()
-
-    @abstractmethod
-    def refresh_collection_summaries(self) -> None:
-        """Make sure that collection summaries for this dataset type are
-        consistent with the contents of the dataset tables.
-        """
 
     datasetType: DatasetType
     """Dataset type whose records this object manages (`DatasetType`).
@@ -667,4 +384,301 @@ class DatasetRecordStorageManager(VersionedExtension):
     @abstractmethod
     def ingest_date_dtype(self) -> type:
         """Return type of the ``ingest_date`` column."""
+        raise NotImplementedError()
+
+    @abstractmethod
+    def insert(
+        self,
+        dataset_type: DatasetType,
+        run: RunRecord,
+        data_ids: Iterable[DataCoordinate],
+        id_generation_mode: DatasetIdGenEnum = DatasetIdGenEnum.UNIQUE,
+    ) -> list[DatasetRef]:
+        """Insert one or more dataset entries into the database.
+
+        Parameters
+        ----------
+        dataset_type : `DatasetType`
+            Type of the datasets to insert.  Storage class is ignored.
+        run : `RunRecord`
+            The record object describing the `~CollectionType.RUN` collection
+            these datasets will be associated with.
+        data_ids : `~collections.abc.Iterable` [ `DataCoordinate` ]
+            Expanded data IDs (`DataCoordinate` instances) for the
+            datasets to be added.   The dimensions of all data IDs must be the
+            same as ``dataset_type.dimensions``.
+        id_generation_mode : `DatasetIdGenEnum`
+            With `UNIQUE` each new dataset is inserted with its new unique ID.
+            With non-`UNIQUE` mode ID is computed from some combination of
+            dataset type, dataId, and run collection name; if the same ID is
+            already in the database then new record is not inserted.
+
+        Returns
+        -------
+        datasets : `list` [ `DatasetRef` ]
+            References to the inserted datasets.
+        """
+        raise NotImplementedError()
+
+    @abstractmethod
+    def import_(
+        self,
+        dataset_type: DatasetType,
+        run: RunRecord,
+        datasets: Iterable[DatasetRef],
+    ) -> list[DatasetRef]:
+        """Insert one or more dataset entries into the database.
+
+        Parameters
+        ----------
+        dataset_type : `DatasetType`
+            Type of all datasets.  Storage class is ignored.
+        run : `RunRecord`
+            The record object describing the `~CollectionType.RUN` collection
+            these datasets will be associated with.
+        datasets : `~collections.abc.Iterable` of `DatasetRef`
+            Datasets to be inserted.  Datasets can specify ``id`` attribute
+            which will be used for inserted datasets. All dataset IDs must
+            have the same type (`int` or `uuid.UUID`), if type of dataset IDs
+            does not match type supported by this class then IDs will be
+            ignored and new IDs will be generated by backend.
+
+        Returns
+        -------
+        datasets : `list` [ `DatasetRef` ]
+            References to the inserted or existing datasets.
+
+        Notes
+        -----
+        The ``datasetType`` and ``run`` attributes of datasets are supposed to
+        be identical across all datasets but this is not checked and it should
+        be enforced by higher level registry code. This method does not need
+        to use those attributes from datasets, only ``dataId`` and ``id`` are
+        relevant.
+        """
+        raise NotImplementedError()
+
+    @abstractmethod
+    def delete(self, datasets: Iterable[DatasetId | DatasetRef]) -> None:
+        """Fully delete the given datasets from the registry.
+
+        Parameters
+        ----------
+        datasets : `~collections.abc.Iterable` [ `DatasetId` or `DatasetRef` ]
+            Datasets to be deleted.  If `DatasetRef` instances are passed,
+            only the `DatasetRef.id` attribute is used.
+        """
+        raise NotImplementedError()
+
+    @abstractmethod
+    def associate(
+        self, dataset_type: DatasetType, collection: CollectionRecord, datasets: Iterable[DatasetRef]
+    ) -> None:
+        """Associate one or more datasets with a collection.
+
+        Parameters
+        ----------
+        dataset_type : `DatasetType`
+            Type of all datasets.
+        collection : `CollectionRecord`
+            The record object describing the collection.  ``collection.type``
+            must be `~CollectionType.TAGGED`.
+        datasets : `~collections.abc.Iterable` [ `DatasetRef` ]
+            Datasets to be associated.  All datasets must be resolved and have
+            the same `DatasetType` as ``dataset_type``.
+
+        Notes
+        -----
+        Associating a dataset into collection that already contains a
+        different dataset with the same `DatasetType` and data ID will remove
+        the existing dataset from that collection.
+
+        Associating the same dataset into a collection multiple times is a
+        no-op, but is still not permitted on read-only databases.
+        """
+        raise NotImplementedError()
+
+    @abstractmethod
+    def disassociate(
+        self, dataset_type: DatasetType, collection: CollectionRecord, datasets: Iterable[DatasetRef]
+    ) -> None:
+        """Remove one or more datasets from a collection.
+
+        Parameters
+        ----------
+        dataset_type : `DatasetType`
+            Type of all datasets.
+        collection : `CollectionRecord`
+            The record object describing the collection.  ``collection.type``
+            must be `~CollectionType.TAGGED`.
+        datasets : `~collections.abc.Iterable` [ `DatasetRef` ]
+            Datasets to be disassociated.  All datasets must be resolved and
+            have the same `DatasetType` as ``self``.
+        """
+        raise NotImplementedError()
+
+    @abstractmethod
+    def certify(
+        self,
+        dataset_type: DatasetType,
+        collection: CollectionRecord,
+        datasets: Iterable[DatasetRef],
+        timespan: Timespan,
+        context: SqlQueryContext,
+    ) -> None:
+        """Associate one or more datasets with a calibration collection and a
+        validity range within it.
+
+        Parameters
+        ----------
+        dataset_type : `DatasetType`
+            Type of all datasets.
+        collection : `CollectionRecord`
+            The record object describing the collection.  ``collection.type``
+            must be `~CollectionType.CALIBRATION`.
+        datasets : `~collections.abc.Iterable` [ `DatasetRef` ]
+            Datasets to be associated.  All datasets must be resolved and have
+            the same `DatasetType` as ``self``.
+        timespan : `Timespan`
+            The validity range for these datasets within the collection.
+        context : `SqlQueryContext`
+            The object that manages database connections, temporary tables and
+            relation engines for this query.
+
+        Raises
+        ------
+        ConflictingDefinitionError
+            Raised if the collection already contains a different dataset with
+            the same `DatasetType` and data ID and an overlapping validity
+            range.
+        CollectionTypeError
+            Raised if
+            ``collection.type is not CollectionType.CALIBRATION`` or if
+            ``self.datasetType.isCalibration() is False``.
+        """
+        raise NotImplementedError()
+
+    @abstractmethod
+    def decertify(
+        self,
+        dataset_type: DatasetType,
+        collection: CollectionRecord,
+        timespan: Timespan,
+        *,
+        data_ids: Iterable[DataCoordinate] | None = None,
+        context: SqlQueryContext,
+    ) -> None:
+        """Remove or adjust datasets to clear a validity range within a
+        calibration collection.
+
+        Parameters
+        ----------
+        dataset_type : `DatasetType`
+            Type of all datasets.
+        collection : `CollectionRecord`
+            The record object describing the collection.  ``collection.type``
+            must be `~CollectionType.CALIBRATION`.
+        timespan : `Timespan`
+            The validity range to remove datasets from within the collection.
+            Datasets that overlap this range but are not contained by it will
+            have their validity ranges adjusted to not overlap it, which may
+            split a single dataset validity range into two.
+        data_ids : `~collections.abc.Iterable` [ `DataCoordinate` ], optional
+            Data IDs that should be decertified within the given validity range
+            If `None`, all data IDs for ``dataset_type`` in ``collection`` will
+            be decertified.
+        context : `SqlQueryContext`
+            The object that manages database connections, temporary tables and
+            relation engines for this query.
+
+        Raises
+        ------
+        CollectionTypeError
+            Raised if ``collection.type is not CollectionType.CALIBRATION``.
+        """
+        raise NotImplementedError()
+
+    @abstractmethod
+    def make_relation(
+        self,
+        dataset_type: DatasetType,
+        *collections: CollectionRecord,
+        columns: Set[str],
+        context: SqlQueryContext,
+    ) -> Relation:
+        """Return a `sql.Relation` that represents a query for for this
+        `DatasetType` in one or more collections.
+
+        Parameters
+        ----------
+        dataset_type : `DatasetType`
+            Type of dataset to query for.
+        *collections : `CollectionRecord`
+            The record object(s) describing the collection(s) to query.  May
+            not be of type `CollectionType.CHAINED`.  If multiple collections
+            are passed, the query will search all of them in an unspecified
+            order, and all collections must have the same type.  Must include
+            at least one collection.
+        columns : `~collections.abc.Set` [ `str` ]
+            Columns to include in the relation.  See `Query.find_datasets` for
+            most options, but this method supports one more:
+
+            - ``rank``: a calculated integer column holding the index of the
+                collection the dataset was found in, within the ``collections``
+                sequence given.
+        context : `SqlQueryContext`
+            The object that manages database connections, temporary tables and
+            relation engines for this query.
+
+        Returns
+        -------
+        relation : `~lsst.daf.relation.Relation`
+            Representation of the query.
+        """
+        raise NotImplementedError()
+
+    @abstractmethod
+    def make_query_joiner(
+        self, dataset_type: DatasetType, collections: Sequence[CollectionRecord], fields: Set[str]
+    ) -> QueryJoiner:
+        """Make a `..direct_query_driver.QueryJoiner` that represents a search
+        for datasets of this type.
+
+        Parameters
+        ----------
+        dataset_type : `DatasetType`
+            Type of dataset to query for.
+        collections : `~collections.abc.Sequence` [ `CollectionRecord` ]
+            Collections to search, in order, after filtering out collections
+            with no datasets of this type via collection summaries.
+        fields : `~collections.abc.Set` [ `str` ]
+            Names of fields to make available in the joiner.  Options include:
+
+            - ``dataset_id`` (UUID)
+            - ``run`` (collection name, `str`)
+            - ``collection`` (collection name, `str`)
+            - ``collection_key`` (collection primary key, manager-dependent)
+            - ``timespan`` (validity range, or unbounded for non-calibrations)
+            - ``ingest_date`` (time dataset was ingested into repository)
+
+            Dimension keys for the dataset type's required dimensions are
+            always included.
+
+        Returns
+        -------
+        joiner : `..direct_query_driver.QueryJoiner`
+            A query-construction object representing a table or subquery.
+        """
+        raise NotImplementedError()
+
+    @abstractmethod
+    def refresh_collection_summaries(self, dataset_type: DatasetType) -> None:
+        """Make sure that collection summaries for this dataset type are
+        consistent with the contents of the dataset tables.
+
+        Parameters
+        ----------
+        dataset_type : `DatasetType`
+            Dataset type whose summary entries should be refreshed.
+        """
         raise NotImplementedError()
